requirements.txt
streamlit==1.28.1
openai==1.3.7
pandas==2.0.3
plotly==5.17.0
python-dotenv==1.0.0
